In [eauto] and [typeclasses eauto]: use [eapply] for every [Hint Resolve].
The hints are classified on whether they can be used by [auto] or not according to a heuristic. Sometimes, the heuristic is too optimistic and accepts a hint [h] where [apply h] can fail while [eapply h] succeeds. This resulted in a counter-intuitive situation where [eauto using h] would fail to progress (because it used [apply h]) while [eapply h] solved the goal.