A set of simple apply tactics based on the refinement mechanism.
These are very simple minded apply-s I call [apply_arity]. They are given a head term and its type, and adds the appropriate number of arguments (from "with" bindings if available or new evars otherwise) according the the number of [forall] in the type (it does not do any type transformation). The resulting term is then refined.

These are useful tactics when we know precisely the number of arguments we want. For instance in tactics such as [constructor] or [reflexivity].

There are, as usual, a [apply] and [eapply] variant. I've also added the (more natural in the current implementation) variant I called [full_apply_arity] which generates its evars as dependent subgoals.