Add an α-equivalence test for [constr_expr].
This test is widely incorrect, as the use of unscoped state update to manage the relation between binding names means that, for instance [(fun a => a) a] and [(fun b => b) b] will appear as α-equivalent.

To fix it, it would suffice to modify updates to be scoped, but the [constr_expr] type is not very adapted for this sort of change (because binders and the term they bind into may be far apart (due to patterns and list of binders)). Such a fix would be a serious endeavour, probably reorganising significantly the equality function.

My guess is that we would have to add a third argument to [binder_eq] which is the list of checks [unit->bool] to be made under the scope of the binder.